Reusable custom control onChange
